rename: firstCoupon 위치 이동
